Compile^^
The original code worked on 32-bit only, the previous Compile++ patch worked on
64-bit only. Hopefully this will work on both ;)

+++ tests/searchtest.cpp
@@ -282,7 +282,7 @@ void SearchTest::testHyphenAtEndOfLineWithoutYOverlap() {
         Okular::NormalizedRect(0.0, 0.4, 0.7, 0.5)
     };
 
-    unsigned long n = sizeof(text)/sizeof(QString);
+    size_t n = sizeof(text)/sizeof(QString);
     QCOMPARE(n, sizeof(rect)/sizeof(Okular::NormalizedRect));
 
     Okular::Page* page;
